+++ libs/sdk-core/CHANGELOG.md
@@ -7,10 +7,12 @@
 - Support new endpoints to download activity log data
 - Updated `feedback` method to allow feedback on specific text blocks
 - Add `entities` property to `FieldMetadata` interface
+- Deprecate `extracted` parameter from `/search`, `/find` and `/ask` endpoints
 
 ### Breaking changes
 
 - Remove support for old activity log endpoints: `/feedback`, `/activity/download` and `activity/downloads`
+- Remove `PARAGRAPH/S`, `VECTOR/S` and `DOCUMENT` values from `Search.Features` and `Ask.Features` enums
 
 # 1.22.0 (2024-10-14)